New: Class structure.
