+++ cleantalk/lib/Cleantalk/Antispam/SFW.php
@@ -1,17 +1,14 @@
 <?php
-namespace CleantalkBase;
 
-require_once 'CleantalkDB.php';
-require_once 'CleantalkHelper.php';
-require_once 'CleantalkAPI.php';
+namespace Cleantalk\Antispam;
 
 /**
  * CleanTalk SpamFireWall base class.
  * Compatible with any CMS.
  *
- * @depends       CleantalkHelper class
- * @depends       CleantalkAPI class
- * @depends       CleantalkDB class
+ * @depends       \Cleantalk\Common\Helper class
+ * @depends       \Cleantalk\Common\API class
+ * @depends       \Cleantalk\Common\DB class
  *
  * @version       3.3
  * @author        Cleantalk team (welcome@cleantalk.org)
@@ -19,7 +16,7 @@ require_once 'CleantalkAPI.php';
  * @license       GNU/GPL: http://www.gnu.org/copyleft/gpl.html
  * @see           https://github.com/CleanTalk/php-antispam
  */
-class CleantalkSFW
+class SFW
 {
 	public $ip = 0;
 	
@@ -63,6 +60,9 @@ class CleantalkSFW
 	protected $data_table;
 	protected $log_table;
 	
+	private $api;
+	private $helper;
+	
 	//Debug
 	public $debug;
 	public $debug_data = '';
@@ -71,18 +71,35 @@ class CleantalkSFW
 	 * CleantalkSFW_Base constructor.
 	 * Creates Database driver instance.
 	 */
-	public function __construct()
-	{
-		if(empty($this->db)){
-			// Creating database object. Depends on current CMS.
-			$this->db = CleantalkDB::getInstance();
-			
-			// Use default tables if not specified
-			$this->data_table = defined('CLEANTALK_TBL_FIREWALL_DATA') ? CLEANTALK_TBL_FIREWALL_DATA : $this->db->prefix . 'cleantalk_sfw';
-			$this->log_table  = defined('CLEANTALK_TBL_FIREWALL_LOG')  ? CLEANTALK_TBL_FIREWALL_LOG  : $this->db->prefix . 'cleantalk_sfw_logs';
-		}
+	public function __construct($db = null, $api = null, $helper = null){
+		
+		// // Creating database object. Depends on current CMS.
+		$this->db     = empty($db)     ? \Cleantalk\Common\DB::getInstance()     : $db;
+		$this->helper = empty($helper) ? \Cleantalk\Common\Helper::getInstance() : $helper;
+		$this->api    = empty($api)    ? \Cleantalk\Common\API::getInstance()    : $api;
 		
 		$this->debug = isset($_GET['debug']) && intval($_GET['debug']) === 1 ? true : false;
+	}
+	
+	/**
+	 * @return \Cleantalk\Common\DB::$instance
+	 */
+	protected function db(){
+		return $this->db;
+	}
+	
+	/**
+	 * @return \Cleantalk\Common\API::$instance
+	 */
+	protected function api(){
+		return $this->api;
+	}
+	
+	/**
+	 * @return \Cleantalk\Common\Helper::$instance
+	 */
+	protected function helper(){
+		return $this->helper;
 	}
 	
 	/**
@@ -95,12 +112,12 @@ class CleantalkSFW
 	 */
 	public function ip__get($ips_input = array('real', 'remote_addr', 'x_forwarded_for', 'x_real_ip', 'cloud_flare'), $v4_only = true){
 		
-		$result = CleantalkHelper::ip__get($ips_input, $v4_only);
+		$result = $this->helper()->ip__get($ips_input, $v4_only);
 		
 		$result = !empty($result) ? array('real' => $result) : array();
 		
 		if(isset($_GET['sfw_test_ip'])){
-			if(CleantalkHelper::ip__validate($_GET['sfw_test_ip']) !== false){
+			if($this->helper()->ip__validate($_GET['sfw_test_ip']) !== false){
 				$result['sfw_test'] = $_GET['sfw_test_ip'];
 				$this->test = true;
 			}
@@ -129,12 +146,12 @@ class CleantalkSFW
 				$this->blocked_ips[$origin] = array(
 					'ip'      => $current_ip,
 					'network' => long2ip($this->db->result['network']),
-					'mask'    => CleantalkHelper::ip__mask__long_to_number($this->db->result['mask']),
+					'mask'    => $this->helper()->ip__mask__long_to_number($this->db->result['mask']),
 				);
 				$this->all_ips[$origin] = array(
 					'ip'      => $current_ip,
 					'network' => long2ip($this->db->result['network']),
-					'mask'    => CleantalkHelper::ip__mask__long_to_number($this->db->result['mask']),
+					'mask'    => $this->helper()->ip__mask__long_to_number($this->db->result['mask']),
 					'status'  => -1,
 				);
 			}else{
@@ -203,7 +220,7 @@ class CleantalkSFW
 			unset($key, $value);
 			
 			//Sending the request
-			$result = CleantalkAPI::method__sfw_logs($ct_key, $data);
+			$result = $this->api()->method__sfw_logs($ct_key, $data);
 			
 			//Checking answer and deleting all lines from the table
 			if(empty($result['error'])){
@@ -237,8 +254,8 @@ class CleantalkSFW
 			
 			sleep(6);
 			
-			$result = CleantalkAPI::method__get_2s_blacklists_db($ct_key, 'file');
-						
+			$result = $this->api()->method__get_2s_blacklists_db($ct_key, 'file');
+			
 			if(empty($result['error'])){
 			
 				if( !empty($result['file_url']) ){
@@ -247,7 +264,7 @@ class CleantalkSFW
 					$pattenrs[] = 'get';
 					if(!$immediate) $pattenrs[] = 'async';
 					
-					return CleantalkHelper::http__request(
+					return $this->helper()->http__request(
 						get_option('siteurl'), 
 						array(
 							'spbc_remote_call_token'  => md5($ct_key),
@@ -264,7 +281,7 @@ class CleantalkSFW
 				return $result;
 		}else{
 						
-			if(CleantalkHelper::http__request($file_url, array(), 'get_code') === 200){ // Check if it's there
+			if($this->helper()->http__request($file_url, array(), 'get_code') === 200){ // Check if it's there
 				
 				if(ini_get('allow_url_fopen')){
