+++ cleantalk/lib/Cleantalk/ApbctUni/API.php
@@ -1,10 +1,12 @@
 <?php
 
+namespace Cleantalk\ApbctUni;
+
 /**
  * Class CleantalkAPI.
  * Compatible only with Wordpress.
  *
- * @depends       CleantalkBase\CleantalkAPI
+ * @depends       \Cleantalk\Common\API
  * 
  * @version       1.0
  * @author        Cleantalk team (welcome@cleantalk.org)
@@ -12,4 +14,6 @@
  * @license       GNU/GPL: http://www.gnu.org/copyleft/gpl.html
  * @see           https://github.com/CleanTalk/wordpress-antispam
  */
-class CleantalkAPI extends CleantalkBase\CleantalkAPI{}
+class API extends \Cleantalk\Common\API{
+
+}
