+++ cleantalk/lib/Cleantalk/Common/File.php
@@ -1,11 +1,18 @@
 <?php
 
-namespace Cleantalk;
+namespace Cleantalk\Common;
 
+/**
+ * Class File
+ * Gather functions that works with files.
+ * All methods are static.
+ *
+ * @package Cleantalk
+ */
 class File{
 	
 	/**
-	 * Delete content from file in tag
+	 * Removes content from file in tag
 	 * Tags example:
 	 * //Cleantalk/TAG_NAME/start
 	 * //Cleantalk/TAG_NAME/end
@@ -13,19 +20,35 @@ class File{
 	 * @param string $file_path
 	 * @param string $tag Tag name (Not whole tag. Only tag name.)
 	 *
-	 * @return bool|\Cleantalk\Err
+	 * @return bool|Err
 	 */
-	function clean__tag( $file_path, $tag ){
-		$pattern =  self::tag__php__start( $tag ) . '[\S\s]*?' . self::tag__php__end( $tag );
-		$pattern = \CleantalkHelper::convert_to_regexp( $pattern );
-		return \CleantalkHelper::is_regexp( $pattern )
+	public static function clean__tag( $file_path, $tag ){
+		$pattern = '\s*' . self::tag__php__start( $tag ) . '[\S\s]*?' . self::tag__php__end( $tag );
+		$pattern = \Cleantalk\Common\Helper::convert_to_regexp( $pattern );
+		return \Cleantalk\Common\Helper::is_regexp( $pattern )
+			? self::clean__pattern( $file_path, $pattern )
+			: Err::add( __CLASS__, __FUNCTION__, 'Pattern wrong', $pattern );
+	}
+	
+	/**
+	 * Removes variable from file
+	 *
+	 * @param string $file_path
+	 * @param string $variable Variable name (without "$")
+	 *
+	 * @return bool|Err
+	 */
+	public static function clean__variable( $file_path, $variable ){
+		$pattern = '\s*$' . $variable . '\s?=[\S\s]*?;';
+		$pattern = \Cleantalk\Common\Helper::convert_to_regexp( $pattern );
+		return \Cleantalk\Common\Helper::is_regexp( $pattern )
 			? self::clean__pattern( $file_path, $pattern )
 			: Err::add( __CLASS__, __FUNCTION__, 'Pattern wrong', $pattern );
 	}
 	
 	/**
 	 * Delete given pattern from file.
-	 Works only for first collision.
+	Works only for first collision.
 	 *
 	 * @param string $file_path
 	 * @param string $pattern RegExp
@@ -57,12 +80,52 @@ class File{
 			return Err::add(__CLASS__, __FUNCTION__, 'No file'); // No template PHP file
 	}
 	
+	public static function replace__variable( $file_path, $variable, $value ){
+		$injection = "\n\t\$$variable = " . var_export( $value, true ) . ";";
+		$needle = '\s*\$' . $variable . '\s?=[\S\s]*?;';
+		static::replace__code( $file_path, $injection, $needle );
+	}
+	
+	public static function replace__code( $file_path, $injection, $needle ){
+		
+		if( is_file( $file_path ) ){
+			
+			if( is_writable( $file_path ) ){
+				
+				$file_content = file_get_contents( $file_path );
+				
+				if( $file_content ){
+					
+					$new_content = preg_replace("/$needle/", $injection, $file_content, 1);
+					$result = $new_content !== null ? true : false;
+					
+					if($result){
+						if( file_put_contents( $file_path, $new_content ) ){
+							return true;
+						}else
+							return Err::add(__CLASS__, __FUNCTION__, 'Write error'); // Cannot write new content to template PHP file
+					}else
+						return Err::add(__CLASS__, __FUNCTION__, 'Replacement fail'); // Can't read from template PHP file
+				}else
+					return Err::add(__CLASS__, __FUNCTION__, 'Read fail'); // Can't read from template PHP file
+			}else
+				return Err::add(__CLASS__, __FUNCTION__, 'No right to write in file'); // No PHP file
+		}else
+			return Err::add(__CLASS__, __FUNCTION__, 'File not found', $file_path); // No PHP file
+	}
+	
 	public static function inject__tag__start( $file_path, $tag ){
 		self::inject__code( $file_path, self::tag__php__start( $tag ) );
 	}
 	
 	public static function inject__tag__end( $file_path, $tag ){
 		self::inject__code( $file_path, self::tag__php__end( $tag ) );
+	}
+	
+	public static function inject__variable( $file_path, $variable, $value, $compact = false ){
+		$value = var_export( $value, true );
+		$value = $compact ? preg_replace( '/\s*/', '', $value ) : $value;
+		self::inject__code( $file_path, "\$$variable = $value;" );
 	}
 	
 	public static function inject__code( $file_path, $injection, $needle = '<\?php', $tag = null ){
@@ -103,7 +166,7 @@ class File{
 			}else
 				return Err::add(__CLASS__, __FUNCTION__, 'No right to write in file'); // No PHP file
 		}else
-			return Err::add(__CLASS__, __FUNCTION__, 'File not found'); // No PHP file
+			return Err::add(__CLASS__, __FUNCTION__, 'File not found', $file_path); // No PHP file
 	}
 	
 	public static function tag__php__start( $tag ){
