+++ cleantalk/lib/Cleantalk/Variables/Request.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Cleantalk\Variables;
+
+/**
+ * Class Request
+ * Safety handler for $_REQUEST
+ *
+ * @usage \Cleantalk\Variables\Request::get( $name );
+ *
+ * @package Cleantalk\Variables
+ */
+class Request extends SuperGlobalVariables{
+	
+	static $instance;
+	
+	/**
+	 * Gets given $_REQUEST variable and seva it to memory
+	 * @param $name
+	 *
+	 * @return mixed|string
+	 */
+	protected function get_variable( $name ){
+		
+		// Return from memory. From $this->variables
+		if(isset(static::$instance->variables[$name]))
+			return static::$instance->variables[$name];
+		
+		$value = isset( $_REQUEST[ $name ] ) ? $_REQUEST[ $name ]	: '';
+		
+		// Remember for thurther calls
+		static::getInstance()->remember_variable( $name, $value );
+		
+		return $value;
+	}
+}
