+++ cleantalk/lib/autoloader.php
@@ -3,7 +3,7 @@
 spl_autoload_register( 'ct_autoloader' );
 
 /**
- * Autoloader
+ * Autoloader for \Cleantalk\* classes
  *
  * @param string $class
  *
@@ -13,7 +13,7 @@ function ct_autoloader( $class ){
 	
 	// Register class auto loader
 	// Custom modules
-	if( strpos( $class, 'Cleantalk' ) !== false ) {
+	if( strpos( $class, 'Cleantalk' ) !== false ){
 		$class_file = __DIR__ . DIRECTORY_SEPARATOR . $class . '.php';
 		if( file_exists( $class_file ) ){
 			require_once( $class_file );
